Multiplayer seems to work. Difficult to test on one system.
I can see both sets of controllers, and move one around, but, since they are both on the same machine (one running on windows, and one in the Unity Editor), it moves the wrong person. I can look at the alternate version being moved around by my inputs, but cannot switch to that version. More testing needs to be done.